# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Kampungku Housing Website API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /settings/homepage:
    get:
      operationId: getHomepageSettings
      tags: [settings]
      summary: Get homepage settings
      responses:
        "200":
          description: Homepage settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomepageSettings"
    put:
      operationId: updateHomepageSettings
      tags: [settings]
      summary: Update homepage settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HomepageSettingsBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomepageSettings"
  /settings/contact:
    get:
      operationId: getContactSettings
      tags: [settings]
      summary: Get contact page settings
      responses:
        "200":
          description: Contact settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSettings"
    put:
      operationId: updateContactSettings
      tags: [settings]
      summary: Update contact page settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactSettingsBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSettings"
  /settings/email:
    get:
      operationId: getEmailSettings
      tags: [settings]
      summary: Get email settings
      responses:
        "200":
          description: Email settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailSettings"
    put:
      operationId: updateEmailSettings
      tags: [settings]
      summary: Update email settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailSettingsBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailSettings"
...
  /settings/drive:
    get:
      operationId: getDriveSettings
      tags: [settings]
      summary: Get Google Drive settings
      responses:
        "200":
          description: Drive settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriveSettings"
    put:
      operationId: updateDriveSettings
      tags: [settings]
      summary: Update Google Drive settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriveSettingsBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriveSettings"
...
  /settings/cctv:
    get:
      operationId: getCctvSettings
      tags: [settings]
      summary: Get CCTV settings
      responses:
        "200":
          description: CCTV settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CctvSettings"
    put:
      operationId: updateCctvSettings
      tags: [settings]
      summary: Update CCTV settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CctvSettingsBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CctvSettings"
  /settings/documents-page:
    get:
      operationId: getDocumentsPageSettings
      tags: [settings]
      summary: Get documents page settings
      responses:
        "200":
          description: Documents page settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentsPageSettings"
    put:
      operationId: updateDocumentsPageSettings
      tags: [settings]
      summary: Update documents page settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentsPageSettingsBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentsPageSettings"
...
components:
  schemas:
...
    HomepageSettings:
      type: object
      properties:
        id:
          type: integer
        heroTitle:
          type: string
        heroSubtitle:
          type: string
        heroImageUrl:
          type: string
        heroBanners:
          type: array
          items:
            type: string
        aboutTitle:
          type: string
        aboutText:
          type: string
        aboutImageUrl:
          type: string
        featuresTitle:
          type: string
        features:
          type: array
          items:
            $ref: "#/components/schemas/Feature"
        statsItems:
          type: array
          items:
            $ref: "#/components/schemas/StatItem"
        updatedAt:
          type: string
      required:
        - id
        - heroTitle
        - heroSubtitle
    HomepageSettingsBody:
      type: object
      properties:
        heroTitle:
          type: string
        heroSubtitle:
          type: string
        heroImageUrl:
          type: string
        heroBanners:
          type: array
          items:
            type: string
        aboutTitle:
          type: string
        aboutText:
          type: string
        aboutImageUrl:
          type: string
        featuresTitle:
          type: string
        features:
          type: array
          items:
            $ref: "#/components/schemas/Feature"
        statsItems:
          type: array
          items:
            $ref: "#/components/schemas/StatItem"
    Feature:
      type: object
      properties:
        icon:
          type: string
        title:
          type: string
        description:
          type: string
      required:
        - title
        - description
    StatItem:
      type: object
      properties:
        value:
          type: string
        label:
          type: string
      required:
        - value
        - label
    ContactSettings:
      type: object
      properties:
        id:
          type: integer
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        mapEmbedUrl:
          type: string
        mapLat:
          type: number
        mapLng:
          type: number
        officeHours:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - address
    ContactSettingsBody:
      type: object
      properties:
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        mapEmbedUrl:
          type: string
        mapLat:
          type: number
        mapLng:
          type: number
        officeHours:
          type: string
    EmailSettings:
      type: object
      properties:
        id:
          type: integer
        gmailUser:
          type: string
        gmailAppPassword:
          type: string
        recipientEmail:
          type: string
        senderName:
          type: string
        updatedAt:
          type: string
      required:
        - id
    EmailSettingsBody:
      type: object
      properties:
        gmailUser:
          type: string
        gmailAppPassword:
          type: string
        recipientEmail:
          type: string
        senderName:
          type: string
    DriveSettings:
      type: object
      properties:
        id:
          type: integer
        appsScriptUrl:
          type: string
        folderId:
          type: string
        updatedAt:
          type: string
      required:
        - id
    DriveSettingsBody:
      type: object
      properties:
        appsScriptUrl:
          type: string
        folderId:
          type: string
    CctvSettings:
      type: object
      properties:
        id:
          type: integer
        pagePassword:
          type: string
        pageTitle:
          type: string
        pageDescription:
          type: string
        updatedAt:
          type: string
      required:
        - id
    CctvSettingsBody:
      type: object
      properties:
        pagePassword:
          type: string
        pageTitle:
          type: string
        pageDescription:
          type: string
    DocumentsPageSettings:
      type: object
      properties:
        id:
          type: integer
        pageTitle:
          type: string
        pageDescription:
          type: string
        updatedAt:
          type: string
      required:
        - id
    DocumentsPageSettingsBody:
      type: object
      properties:
        pageTitle:
          type: string
        pageDescription:
          type: string
...