# lib/api-zod/src/generated/types/cctvSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kampungku Housing Website API
 * OpenAPI spec version: 0.1.0
 */

export interface CctvSettings {
  id: number;
  pagePassword?: string;
  pageTitle?: string;
  pageDescription?: string;
  updatedAt?: string;
}