# lib/api-zod/src/generated/types/contactSettingsBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kampungku Housing Website API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactSettingsBody {
  address?: string;
  phone?: string;
  email?: string;
  mapEmbedUrl?: string;
  mapLat?: number;
  mapLng?: number;
  officeHours?: string;
}