# lib/api-zod/src/generated/types/homepageSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kampungku Housing Website API
 * OpenAPI spec version: 0.1.0
 */
import type { Feature } from "./feature";
import type { StatItem } from "./statItem";

export interface HomepageSettings {
  id: number;
  heroTitle: string;
  heroSubtitle: string;
  heroImageUrl?: string;
  heroBanners?: string[];
  aboutTitle?: string;
  aboutText?: string;
  aboutImageUrl?: string;
  featuresTitle?: string;
  features?: Feature[];
  statsItems?: StatItem[];
  updatedAt?: string;
}